Rename hook folder
